engine: split router.rs into per-domain modules
router.rs was a 3,423-line file dominated by one 231-arm `match
req.method.as_str()` in async fn route. Same shape as the
god-object pattern from the blog post — adding an RPC, even a
simple read-only one, meant scrolling past 2,400 lines of
unrelated arms to find your domain's section.

Split per domain (no behavior change, no API change):

  router/mod.rs       1053 lines  -- handle_rpc_request entry
                                      point, auth/role gating,
                                      audit hooks, prefix-dispatch,
                                      shared helpers (ok/err/parse_params/
                                      require_str/check_*) and a
                                      handful of cross-cutting fns
                                      (vm_snapshot/vm_clone, etc.)
                                      that touch multiple services.
  router/auth.rs       171   12 arms
  router/audit.rs       31    1 arm
  router/alerts.rs      68    6 arms (alert.*, telemetry.*,
                                      system.alerts)
  router/system.rs     599   62 arms (system.*, firmware.*)
  router/service.rs    125    7 arms
  router/notifications  42    3 arms
  router/backup.rs      91   10 arms
  router/fs.rs         272   28 arms (fs.*, device.*)
  router/bcachefs.rs    43    3 arms
  router/subvolume.rs  283   14 arms
  router/snapshot.rs    64    4 arms
  router/share.rs      402   28 arms (share.nfs/smb/iscsi/nvmeof)
  router/smb.rs        119    9 arms (smb.user/smb.group)
  router/vm.rs          98   13 arms
  router/apps.rs       239   31 arms

Each domain module exposes `pub(super) async fn try_route(req,
state, session) -> Option<Response>`. Inside is the same `match
req.method.as_str()` shape, just scoped to that domain, ending in
`_ => return None` so unknown methods fall through.

The new top-level `route()` in mod.rs is a 20-line prefix dispatcher
(`match req.method.split_once('.').0`) — adding a new RPC to an
existing domain now stops touching the other 14 modules. Two
special cases: `system.alerts` lives in alerts.rs (so the `system`
prefix has a one-line check for it before falling through to
system.rs), and `firmware.*` lives in system.rs (along with the
rest of update-related RPCs).

Helpers shared across domains (ok/err/invalid/parse_params/
require_str/require_protocol/str_param/check_block_device_conflict
/check_subvolume_in_use/check_filesystem_in_use/vm_snapshot/
vm_clone/list_vm_images/ensure_images_subvolume/fetch_metrics_json/
read_bcachefs_error_count) became `pub(super)` and stayed in mod.rs.
They can move into their natural home in a follow-up; the win this
PR ships is the 231-arm match being gone.

Build + clippy --all-targets -- -D warnings + cargo test all green.
