+++ engine/nasty-engine/src/router/bcachefs.rs
@@ -1,0 +1,43 @@
+//! RPC arms in the `bcachefs.*` domain. Extracted from the historical
+//! 231-arm `match` in `router.rs`. Returns `Some(response)` when the
+//! method matches, `None` when it falls through to another domain.
+
+#![allow(unused_imports, unused_variables)]
+
+use nasty_common::{ErrorCode, Request, Response};
+use serde::Deserialize;
+
+use super::*;
+use crate::AppState;
+use crate::auth::{Role, Session};
+
+pub(super) async fn try_route(
+    req: &Request,
+    state: &AppState,
+    session: &Session,
+) -> Option<Response> {
+    Some(match req.method.as_str() {
+        "bcachefs.usage" => match require_str(req, "name") {
+            Ok(name) => match state.filesystems.bcachefs_usage(name).await {
+                Ok(v) => ok(req, v),
+                Err(e) => err(req, e),
+            },
+            Err(r) => r,
+        },
+        "bcachefs.top" => match require_str(req, "name") {
+            Ok(name) => match state.filesystems.bcachefs_top(name).await {
+                Ok(v) => ok(req, v),
+                Err(e) => err(req, e),
+            },
+            Err(r) => r,
+        },
+        "bcachefs.timestats" => match require_str(req, "name") {
+            Ok(name) => match state.filesystems.bcachefs_timestats(name).await {
+                Ok(v) => ok(req, v),
+                Err(e) => err(req, e),
+            },
+            Err(r) => r,
+        },
+        _ => return None,
+    })
+}
